+++ lib/karafka/schemas/consumer_group.rb
@@ -24,7 +24,6 @@ module Karafka
       required(:offset_commit_threshold).filled(:int?)
       required(:offset_retention_time) { none?.not > int? }
       required(:heartbeat_interval).filled(:int?, gteq?: 0)
-      required(:topic_mapper).filled
       required(:connect_timeout).filled(:int?, gt?: 0)
       required(:socket_timeout).filled(:int?, gt?: 0)
       required(:max_wait_time).filled(:int?, gteq?: 0)